+++ StaticApp/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
-from StaticApp import views
+from StaticApp import views2
 
 urlpatterns = [
-    path(r'',views.index),
+    path(r'',views2.index),
 ]
