+++ StaticApp/views2.py
