+++ TemplateApp/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
-from TemplateApp import views
+from TemplateApp import views1
 
 urlpatterns = [
-    path(r'',views.index),
+    path(r'',views1.index),
 ]
